Move webpackImporter to dedicated module
